Move some icaltime functions to private API
Moved to icaltime_p.[c,h]:
* The icaltime_t type
* The icaltime_span type
* icaltime_span_new
* icaltime_span_overlaps
* icaltime_span_contains
